fix: self-host libraries
Instead of requiring the user to retrieve Font Awesome and HTMX from a separate
CDN, we instead bundle the files during the build.

This also adds a script that reduces the size of the font files to only the
glyphs we need.

Also added is a `Makefile` that takes care of performing the steps now required
for a build.