feat: add tracking using Umami
To get a rough feeling of the organic traffic to the site, this commit adds an
integration to a self-hosted instance of Umami. Umami is a GDPR-compliant,
lightweight tracking tool that literally only tracks page-views with a tiny bit
of metadata. This should be just enough to be useful/interesting, but not so
much that it could be considered an invasion of privacy by a visitor.

To not require any cross-domain requests, two functions are added that simply
forward the incoming requests to the Umami instance.